sched: fastpath for prev_cpu
If the prev_cpu can accommodate the task and is idle we should
prefer it instead of going through all the groups and their cpus.

This significantly helps with performance. Especially in
situations where tasks sleeps and wakes up often - the prev_cpu
is likely left idle and we can avail of the hot caches.

This also makes bias_to_prev_cpu() unnecessary, remove it.

Change-Id: If7dc2081b9178f0025545e03db8a33a1f43e32a0
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
[Fixup merge errors]
Signed-off-by: Zachariah Kennedy <zkennedy87@gmail.com>
Signed-off-by: MadeOfGreat <ravenklawasd@gmail.com>
Signed-off-by: rk134 <rahul-k@bigdi.cc>