drivers: camera: HACK: Always power down actuator.
After some test with CTS the mm-qcamera-daemon there are
issues with crashes them. In some scenarios the camera subsystem
is restarted. After that the state of the actuator is set to
ACT_DISABLE_STATE which doesn't match the reality.
The interims solution is to power down the actuator in any cases -
it doesn't seem to have any side effects. If not, the regulator
isn't able to switch into an low power mode and leaks battery power.

Note: This has to be replaced with an proper solution.
Signed-off-by: Lau <laststandrighthere@gmail.com>
Signed-off-by: GhostMaster69-dev <rathore6375@gmail.com>
Signed-off-by: rk134 <rahul-k@bigdi.cc>